cropping

+++ tests_multiobject/trackers/eval_sam2.py
@@ -15,7 +15,7 @@ from tqdm import tqdm
 
 from hq_sam.sam_hq.segment_anything import sam_model_registry, SamPredictor
 
-torch.autocast(device_type="cuda", dtype=torch.float16).__enter__()
+torch.autocast(device_type="cuda", dtype=torch.bfloat16).__enter__()
 
 if torch.cuda.get_device_properties(0).major >= 8:
     # turn on tfloat32 for Ampere GPUs (https://pytorch.org/docs/stable/notes/cuda.html#tensorfloat-32-tf32-on-ampere-devices)
